Extract GhosttySurfaceViewDelegate into its own file (length budget)
GhosttySurfaceView.swift exceeded its swift-file-length budget after the
reportID additions; move the delegate protocol + default-impl extension out
(pure move, no visibility changes) so the file lands back under budget.
